src/test/java/org/alan/algorithm/practice/baekjoon/problem/BJ3986Test.java
package org.alan.algorithm.practice.baekjoon.problem;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

public class BJ3986Test {
    static Stream<Arguments> parameters() {
        return Stream.of(
                Arguments.of("""
                        3
                        ABAB
                        AABB
                        ABBA""", "2"),
                Arguments.of("""
                        3
                        AAA
                        AA
                        AB""", "1"),
                Arguments.of("""
                        1
                        ABBABB""", "1")
        );
    }
    
    @ParameterizedTest
    @MethodSource("parameters")
    public void test1(String input, String expected) {
        Assertions.assertEquals(expected, new BJ3986().solve(input));
    }
}
